Updates/fixes from feedback
