+++ frontend/src/concepts/areas/types.ts
@@ -65,7 +65,7 @@ export enum SupportedArea {
   YAML_VIEWER = 'yaml-viewer',
   VLLM_ON_MAAS = 'vllm-on-maas',
   LLMD_GATEWAY_FIELD = 'llmd-gateway-field',
-  API_KEYS = 'my-subscriptions',
+  MY_SUBSCRIPTIONS = 'my-subscriptions',
 
   /* Distributed Workloads areas */
   DISTRIBUTED_WORKLOADS = 'distributed-workloads',
